requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.13.4
certifi==2025.7.14
charset-normalizer==3.4.2
coreapi==2.3.3
coreschema==0.0.4
dj-rest-auth==7.0.1
Django==4.2.23
django-allauth==65.10.0
django-cors-headers==4.4.0
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
MarkupSafe==2.1.5
openai-whisper @ git+https://github.com/openai/whisper.git@c0d2f624c09dc18e709e37c2ad90c039a4eb72a2
openapi-codec==1.3.2
packaging==25.0
phonenumberslite==9.0.10
pydub==0.25.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
simplejson==3.20.1
six==1.17.0
soupsieve==2.7
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.2.3
gunicorn
whitenoise
psycopg2-binary
